Split test kinds, common fixtures, utilities in wdspec tests; r=ato+446296
MozReview-Commit-ID: H2vSkhlfwE9

Upstreamed from https://bugzilla.mozilla.org/show_bug.cgi?id=1337133
