+++ webdriver/actions/mouse.py
@@ -1,0 +1,4 @@
+import pytest
+
+def test_nothing():
+    assert True
