+++ webdriver/actions/support/test_actions_wdspec.html
@@ -1,7 +1,8 @@
+
 <!doctype html>
 <meta charset=utf-8>
 <head>
-    <title>Test Keys</title>
+    <title>Test Actions</title>
     <script>
         var allEvents = {events: []};
         function displayMessage(message) {
